be/requirements.txt
psycopg2-binary
Django >= 2.0
djangorestframework
drf-yasg
django-oauth-toolkit